Move storage engine hooks from store to storage engine
