+++ src/GlobalInitializer.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { useStorageEngine } from './store/storageEngineHooks';
+import { useStorageEngine } from './storage/storageEngineHooks';
 import { GlobalConfigParser } from './GlobalConfigParser';
 import { initializeStorageEngine } from './storage/initialize';
