+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { StorageEngineProvider } from './store/storageEngineHooks';
+import { StorageEngineProvider } from './storage/storageEngineHooks';
 import { GlobalInitializer } from './GlobalInitializer';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
